reels-analizer-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
openai
python-dotenv
pydantic